sched: Enable NEXT_BUDDY
[Forwarded from Andrzej Perczak]
Next buddy is like:
thread 1 is running, try to wake 2, wake fails bcs sth,
go back to 1, when 1 is done then poll for next task
but 2 failed so wake it again

Zachariah (RenderBroken), [4/6/20 1:57 AM]
[In reply to Andrzej Perczak]
Hey! You are mostly correct. It is to benefit cache locality
Here is a good summary:

/*
* Pick the next process, keeping these things in mind, in this order:
* 1) keep things fair between processes/task groups
* 2) pick the "next" process, since someone really wants that to run
* 3) pick the "last" process, for cache locality
* 4) do not run the "skip" process, if something else is available
*/

And it isnt really like stopping a thread,
it is more like picking the next sched entity to run next.
I guess it is possible for the se to have ran earlier but it isnt responsible for stopping it.
Just picking the next se to run based on fairness, cache locality, and group scheduling

Signed-off-by: Rasenkai <rasenkai99@gmail.com>
Change-Id: Ifb71c16479e76de3699b32f010e20eeeaf39cb98