drivers: arm-smmu: Power-on SMMU during freeze
There can be a possibility that runtime powet management
has powered-down SMMU. Now, if there is a freeze request
on top of that, we will see SMMU register access errors,
as the SMMU is already powered-down.

Power-on SMMU during the freeze for accessing the SMMU
registers as part of qcom_free_io_pgtable_ops().

Change-Id: I2a00c6b7e519f79180283c9e99dd677edeb426d7
Signed-off-by: Shreyas K K <quic_shrekk@quicinc.com>